refactor: move the shared build config out of config/ (#2883)
config/ was doing two jobs: a workspace directory the running app writes
to, and the home of two committed build configs. #2879 kept them apart
with an allowlist, which is fail-open — a newly committed file there is
invisible to eslint until someone adds a negation, and the rule had to
be repeated in .gitignore and eslint.config.mjs.

Move both files to build-config/, at the same depth so every extends /
import is a plain string substitution. config/ then joins /data/ and
/artifacts/ as a blanket /config/ ignore, with no negations anywhere.

Two call sites a path substitution alone would miss: devWatchIgnore's
WORKSPACE_RUNTIME_ENTRIES excluded `config` because it held real source,
which is no longer true, and the CI hashFiles() cache keys do not error
on a path that stopped existing — they would have gone stale in silence.

Resolved configuration is unchanged: tsc --showConfig over all 34 package
tsconfigs diffs identical, and all 9 package eslint configs hash identical.
