+++ .gitignore
@@ -39,21 +39,10 @@ test-results-live/
 /github/
 /models/
 /plugins/
+/config/
 /.mulmoclaude/
 /.session-token
 /.server-port
-# `config/` can't join the list above wholesale — it is the one workspace dir
-# that also holds committed build config. Ignore its contents and re-admit
-# those two files, so scheduler state / roles / dashboard and friends stay
-# uncommittable. Prettier reads .gitignore, so this also keeps `yarn format`
-# off files the running app rewrites (#2873).
-#
-# Committing a NEW file here takes a negation in two places: below, and in
-# `eslint.config.mjs`'s ignores. `git add` refuses an ignored path outright,
-# so the first half announces itself; the eslint half does not.
-/config/*
-!/config/eslint.packages.mjs
-!/config/tsconfig.packages.json
 
 # V8 heap snapshots written by the dev server on near-OOM (scripts/dev-server.mjs)
 *.heapsnapshot
