+++ package.json
@@ -36,10 +36,10 @@
     "cli": "tsx packages/bridges/cli/src/index.ts",
     "google:auth": "tsx server/services/google/authCli.ts",
     "telegram": "tsx packages/bridges/telegram/src/index.ts",
-    "lint": "yarn dumpi18n && eslint src server test e2e e2e-live packages scripts batch config --cache --cache-location node_modules/.cache/eslint/ --format ./scripts/eslint-formatter-summary.mjs",
-    "lint:summary": "yarn dumpi18n && eslint src server test e2e e2e-live packages scripts batch config --format json | node scripts/lint-summary.mjs",
+    "lint": "yarn dumpi18n && eslint src server test e2e e2e-live packages scripts batch build-config --cache --cache-location node_modules/.cache/eslint/ --format ./scripts/eslint-formatter-summary.mjs",
+    "lint:summary": "yarn dumpi18n && eslint src server test e2e e2e-live packages scripts batch build-config --format json | node scripts/lint-summary.mjs",
     "dumpi18n": "tsx batch/i18n-dump.ts",
-    "format": "prettier --write '{src,server,test,e2e,e2e-live,packages,scripts,batch,config}/**/*.{ts,mts,mjs,json,yaml,vue}'",
+    "format": "prettier --write '{src,server,test,e2e,e2e-live,packages,scripts,batch,build-config}/**/*.{ts,mts,mjs,json,yaml,vue}'",
     "typecheck": "yarn typecheck:vue && yarn typecheck:server && yarn typecheck:test && yarn typecheck:e2e && yarn typecheck:e2e-live && yarn typecheck:packages",
     "typecheck:packages": "yarn workspaces run typecheck",
     "typecheck:vue": "vue-tsc --noEmit",
